Modified to exclude Auth and its services
