--- Server/src/main/java/com/revature/ecommerce/entities/Auth.java
@@ -1,59 +1,0 @@
-package com.revature.ecommerce.entities;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Id;
-
-import java.util.Objects;
-
-public class Auth {
-
-    @Id
-    String email;
-    @Column
-    String password;
-
-    public Auth() {
-    }
-
-    public Auth(String email, String password) {
-        this.email = email;
-        this.password = password;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Auth auth = (Auth) o;
-        return Objects.equals(email, auth.email) && Objects.equals(password, auth.password);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(email, password);
-    }
-
-    @Override
-    public String toString() {
-        return "Auth{" +
-                "email='" + email + '\'' +
-                ", password='" + password + '\'' +
-                '}';
-    }
-}
